修改销售单订单生成

+++ dxt-third-core/src/main/java/com/dxt/third/core/entity/Product.java
@@ -17,7 +17,7 @@ public class Product {
 
     private String productSerial;
 
-    private Long productNumber;
+    private Integer productNumber;
 
     private String remark;
